Убрал проверку токена на публичных запросах
