+++ src/main/java/hh/forest_of_habits/controller/ShareController.java
@@ -9,9 +9,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.DeleteMapping;
 
 import java.util.UUID;
 
@@ -30,25 +27,5 @@ public class ShareController {
     @GetMapping("/tree/{id}")
     TreeFullResponse getTree(@PathVariable Long id) {
         return treeService.getById(id);
-    }
-
-    @PutMapping("/{id}")
-    UUID makeShared(@PathVariable Long id) {
-        return forestService.makeShared(id, true);
-    }
-
-    @PutMapping
-    void makeShared(@RequestParam Long forestId, @RequestParam Long userId) {
-        forestService.makeShared(forestId, userId, true);
-    }
-
-    @DeleteMapping("/{id}")
-    void makeUnshared(@PathVariable Long id) {
-        forestService.makeShared(id, false);
-    }
-
-    @DeleteMapping
-    void makeUnshared(@RequestParam Long forestId, @RequestParam Long userId) {
-        forestService.makeShared(forestId, userId, false);
     }
 }
